refactor: Input -> TextInput으로 이름 변경
